chatIdが無い時に削除するとエラーになるやつ修正

+++ src/main.ts
@@ -139,8 +139,10 @@ async function removeUser(message: Message) {
   })
   const data = await db.findOne({discordId: discordId})
   db.remove({discordId: discordId})
-  const mes: Message = await (<TextChannel>message.guild.channels.get(process.env.CHANNEL_ID)).fetchMessage(data.chatId)
-  mes.delete()
+  if (data.chatId) {
+    const mes: Message = await (<TextChannel>message.guild.channels.get(process.env.CHANNEL_ID)).fetchMessage(data.chatId)
+    mes.delete()
+  }
 
   channel.send(`<@${discordId}> 登録を消したで`)
 }
